😺[CHORE] 모든 style => style 폴더로 통합, page는 page 폴더로 통합, 사용하지 않는 코드 삭제
