+++ package.json
@@ -10,7 +10,6 @@
     "@testing-library/react": "^13.4.0",
     "@testing-library/user-event": "^13.5.0",
     "@types/three": "^0.149.0",
-    "framer-motion": "^9.0.2",
     "gsap": "^3.11.4",
     "html2canvas": "^1.4.1",
     "react": "^18.2.0",
