+++ src/Component/Slide.js
@@ -1,7 +1,8 @@
-import React, { useEffect, useState, useRef } from "react";
+import React, { useEffect, useRef } from "react";
 import "slick-carousel/slick/slick.css";
 import "slick-carousel/slick/slick-theme.css";
 import Slider from "react-slick";
+import { SliderBody } from "../styles/slideStyles";
 
 const Slide = ({ photos, pickEvent = null, side }) => {
   const settings = {
@@ -25,63 +26,27 @@ const Slide = ({ photos, pickEvent = null, side }) => {
   }, [side]);
 
   return (
-    <>
-      <div className={`slide-body`}>
-        <section className={`slide-section`}>
-          <Slider {...settings} ref={sliderRef}>
-            {photos.map((photo, index) => {
-              return (
-                <div
-                  className={`photo-box`}
-                  style={{
-                    cursor: "pointer",
-                  }}
-                  onClick={() => {
-                    pickEvent(index);
-                  }}
-                >
-                  {photo}
-                </div>
-              );
-            })}
-          </Slider>
-        </section>
-      </div>
-
-      <style jsx>{`
-        div.slide-body {
-          display: flex;
-          justify-content: center;
-          height: 100%;
-        }
-        section.slide-section {
-          display: flex;
-          width: 100%;
-          justify-content: center;
-          align-items: center;
-          overflow-x: hidden;
-        }
-      `}</style>
-      <style jsx global>{`
-        section.slide-section .slick-slider {
-          width: 80%;
-          height: 80%;
-          display: flex;
-          align-items: center;
-          justify-content: center;
-        }
-        .slick-list {
-          box-shadow: rgba(0, 0, 0, 0.24) 0px 3px 8px;
-        }
-        .slick-prev:before,
-        .slick-next:before {
-          color: black;
-        }
-        .photo-box {
-          display: flex !important;
-        }
-      `}</style>
-    </>
+    <SliderBody>
+      <section className={`slide-section`}>
+        <Slider {...settings} ref={sliderRef}>
+          {photos.map((photo, index) => {
+            return (
+              <div
+                className={`photo-box`}
+                style={{
+                  cursor: "pointer",
+                }}
+                onClick={() => {
+                  pickEvent(index);
+                }}
+              >
+                {photo}
+              </div>
+            );
+          })}
+        </Slider>
+      </section>
+    </SliderBody>
   );
 };
