+++ src/router.js
@@ -1,13 +1,10 @@
 import React from "react";
 import { BrowserRouter, Routes, Route } from "react-router-dom";
-import Info from "./Component/Info";
-import Name from "./Component/Name";
-import HomePage from "./Component/HomePage";
-import SelectTamplate from "./Component/SelectTamplate";
-import BasicCardOne from "./Component/business-card/BasicCardOne";
-import BasicCardTwo from "./Component/business-card/BasicCardTwo";
-import BasicCardThree from "./Component/business-card/BasicCardThree";
-import Share from "./Component/Share";
+import Info from "./page/InfoPage";
+import Name from "./page/NamePage";
+import HomePage from "./page/HomePage";
+import SelectTamplatePage from "./page/SelectTamplatePage";
+import Share from "./page/SharePage";
 import ResultPage from "./Component/ResultPage";
 
 function Router() {
@@ -17,9 +14,9 @@ function Router() {
         <Route path="/" element={<HomePage />} />
         <Route path="/name" element={<Name />} />
         <Route path="/info" element={<Info />} />
-        <Route path="/select-template" element={<SelectTamplate />} />
-        <Route path="/share" element={<Share />} />
+        <Route path="/select-template" element={<SelectTamplatePage />} />
         <Route path="/result" element={<ResultPage />} />
+        <Route path="/share" element={<Share />} />
       </Routes>
     </BrowserRouter>
   );
