--- src/styles/global.ts
@@ -1,11 +1,0 @@
-// import { createGlobalStyle } from 'styled-components';
-// import HUFSBA from '../fonts/HUFSBA.otf';
-
-// const GlobalStyle = createGlobalStyle`
-//     @font-face {
-//         font-family: 'HUFSBA';
-//         src: local('HUFSBA'), local('HUFSBA');
-//         font-style: bold;
-//         src: url(${HUFSBA}) format('truetype');
-//     }
-// `;
